requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.1
flask-swagger-ui==4.11.1
PyMySQL==1.1.0
cryptography==42.0.5
flask-jwt-extended==4.6.0
bcrypt==4.1.2